requirements.txt
flask
google-genai
pydantic
python-dotenv
pillow